Remove MonoDrawing
