+++ OpenKh.Game/OpenKh.Game.csproj
@@ -19,7 +19,6 @@
   <ItemGroup>
     <ProjectReference Include="..\OpenKh.Engine\OpenKh.Engine.csproj" />
     <ProjectReference Include="..\OpenKh.Kh2\OpenKh.Kh2.csproj" />
-    <ProjectReference Include="..\XeEngine.Tools.Public\Xe.Drawing\Xe.Drawing.csproj" />
   </ItemGroup>
 
   <ItemGroup>
